+++ src/python/dxpy/utils/describe.py
@@ -415,7 +415,7 @@ def print_project_desc(desc, verbose=False):
         'containsPHI', 'databaseUIViewOnly', 'externalUploadRestricted', 'region', 'storageCost', 'pendingTransfer',
         'atSpendingLimit', 'currentMonthComputeAvailableBudget', 'currentMonthEgressBytesAvailableBudget',
         'currentMonthStorageAvailableBudget', 'currentMonthComputeUsage', 'currentMonthEgressBytesUsage', 
-        'currentMonthExpectedStorageUsage',
+        'currentMonthExpectedStorageUsage', 'defaultSymlink'
         # Following are app container-specific
         'destroyAt', 'project', 'type', 'app', 'appName'
     ]
@@ -453,6 +453,8 @@ def print_project_desc(desc, verbose=False):
         print_json_field('Database UI View Only', desc['databaseUIViewOnly'])
     if 'externalUploadRestricted' in desc and desc['externalUploadRestricted']:
         print_json_field('External Upload Restricted', desc['externalUploadRestricted'])
+    if 'defaultSymlink' in desc and verbose:
+        print_json_field('Default Symlink', desc['defaultSymlink'])
 
     # Usage
     print_field("Created", render_timestamp(desc['created']))